requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Faker==8.10.3
flake8==3.9.2
isort==5.9.3
kombu==5.1.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.1
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.9.6
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
